+++ client/src/test/java/com/microsoft/durabletask/IntegrationTests.java
@@ -356,7 +356,7 @@ public class IntegrationTests extends IntegrationTestBase {
     }
 
     @Test
-    void continueAsNewWithExternalEvents() throws TimeoutException {
+    void continueAsNewWithExternalEvents() throws TimeoutException, InterruptedException{
         final String orchestratorName = "continueAsNewWithExternalEvents";
         final String eventName = "MyEvent";
         final int expectedEventCount = 10;
@@ -1194,7 +1194,7 @@ public class IntegrationTests extends IntegrationTestBase {
     }
 
     @Test
-    void externalEventThenAccept() throws IOException, InterruptedException, TimeoutException {
+    void externalEventThenAccept() throws InterruptedException, TimeoutException {
         final String orchestratorName = "continueAsNewWithExternalEvents";
         final String eventName = "MyEvent";
         final int expectedEventCount = 10;